Move infinite scroll to widgets.
