+++ src/app/pages/api/brewery/components/index.ts
@@ -1,2 +1,1 @@
 export * from './beers/beers.component';
-export * from './beers/infinitescroll.component'
