--- src/app/widgets/directives/index.ts
@@ -1,1 +1,0 @@
-export * from './scroll.directive';
