+++ src/app/widgets/index.ts
@@ -6,6 +6,7 @@ export * from './utils';
 export * from './link';
 export * from './menu';
 export * from './dropdown';
+export * from './scroll';
 export { DraalDataTableModule } from './datatables/datatables.module';
 export { DraalFormsModule } from './form/form.module';
 export { DraalAlertModule } from './alert/alert.module';
