+++ src/app/widgets/scroll/index.ts
@@ -1,0 +1,2 @@
+export * from './scroll.directive';
+export * from './infinitescroll.component';
