+++ src/app/widgets/scroll/infinitescroll.component.scss
